Moved generative program execution into submodule
